backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Agentic AI Framework (smolagents)
smolagents>=1.22.0
litellm==1.52.15

# LangChain (for RAG components only)
langchain>=0.3.12
langchain-community>=0.3.12
langchain-google-genai>=2.0.5

# Vector Database
qdrant-client==1.12.1

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
trafilatura==2.0.0
lxml==5.3.0

# Google AI SDK (for multi-modal)
google-generativeai==0.8.4

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
